Enhance math tests and add N-dimensional minimization functionality
This commit introduces several improvements and additions to the math testing framework and functionality:

1. **New Test File**:
   - Added `math_Min_NDim_Test.cxx` to implement tests for N-dimensional minimization algorithms, ensuring comprehensive coverage of the new features.

2. **Test Updates**:
   - Updated existing tests in `math_Integ_Comparison_Test.cxx`, `math_Integ_Test.cxx`, `math_Min_Comparison_Test.cxx`, and `math_Min_Test.cxx` to dereference optional values for consistency and correctness.

3. **CMake Modifications**:
   - Updated `FILES.cmake` to include the new N-dimensional minimization test file, ensuring it is part of the build process.

These changes aim to enhance the robustness of the math library by ensuring that new features are thoroughly tested and integrated into the existing framework.

+++ src/FoundationClasses/TKMath/GTests/math_Poly_Comparison_Test.cxx
@@ -55,7 +55,7 @@ namespace
     std::vector<double> aRoots;
     if (theResult.IsDone())
     {
-      for (int i = 0; i < theResult.NbRoots; ++i)
+      for (size_t i = 0; i < theResult.NbRoots; ++i)
       {
         aRoots.push_back(theResult.Roots[i]);
       }
